Demo of design chart for Writer styles
Attempt at user-friendly approach to designing LO Writer templates.
Parameters are presented in tabular form. In this demo, vertical
spacings are re-scaled automatically when the font size is changed. The
spreadsheet could be parsed into a template file, but this has not been
done because a working version would need to be more interactive.